added rps and ld-lvl handling

+++ src/phandler.h
@@ -41,8 +41,8 @@ typedef uint8_t * (*receiver_func)();
 int query_create(const struct servant_t *servant);
 int reply_process(const struct servant_t *servant, const uint8_t *answer,
     const size_t answer_size);
-int perform_query(const struct servant_t *servant, sender_func send_query);
-int recv_reply(const struct servant_t *servant, receiver_func recvr,
+ssize_t perform_query(const struct servant_t *servant, sender_func send_query);
+ssize_t recv_reply(const struct servant_t *servant, receiver_func recvr,
     const servant_type_t type);
 
 #endif /* __QUERY_H__ */
